fix(gdal_feature): Remove "override" specifier from "set" function.

+++ test/test_feature.cpp
@@ -31,7 +31,7 @@ TEST_CASE("exception thrown on oversize unsigned int", "[feature]")
     auto x = static_cast<std::size_t>(std::numeric_limits<std::int64_t>::max()) + 1;
 
     MapFeature mf;
-    CHECK_THROWS_WITH(mf.set("field", x), Catch::StartsWith("Value is too large"));
+    CHECK_THROWS_WITH(mf.set("field", static_cast<uint64_t>(x)), Catch::StartsWith("Value is too large"));
 }
 
 TEST_CASE("exception thrown on oversized value in unsigned int array", "[feature]")
